requirements.txt
python-telegram-bot
pyTelegramBotAPI
requests
gTTS
ffmpeg-python
SpeechRecognition
python-dotenv
